feat: creating a directory for home and archive, linking to a dummy archive page
