+++ package.json
@@ -11,6 +11,7 @@
     "@fortawesome/react-fontawesome": "^0.2.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.28.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
