+++ src/App.tsx
@@ -1,10 +1,20 @@
 import React from "react";
-import "./App.css";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
-import { Home } from "./views/Home";
+import './App.css';
+
+import { Home } from './views/home/Home';
+import { Archive } from "./views/archive/Archive";
 
 function App() {
-  return <Home />;
+    return (
+        <Router>
+            <Routes>
+                <Route path="/daily" element={<Home />}/>
+                <Route path="/daily/archive" element={<Archive />}/>
+            </Routes>
+        </Router>
+    );
 }
 
 export default App;
