+++ src/views/archive/Archive.tsx
@@ -1,0 +1,1 @@
+export const Archive = () => <p>Archive!</p>
